refactor!: add alias for latest material resources
