+++ src/library/Uno.Material/Styles/Application/v1/MaterialColors.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					x:Class="Uno.Material.MaterialColors">
+					x:Class="Uno.Material.MaterialColorsV1">
 	<!--
 
 	For more information on the color system, consult this page
