+++ src/library/Uno.Material/Styles/Controls/v1/_Resources.xaml
@@ -9,6 +9,6 @@
 
 	<ResourceDictionary.MergedDictionaries>
 		<um:MaterialFonts />
-		<um:MaterialColors />
+		<um:MaterialColorsV1 />
 	</ResourceDictionary.MergedDictionaries>
 </ResourceDictionary>
