+++ src/samples/Uno.Themes.Samples/Uno.Themes.Samples.Shared/App.xaml
@@ -8,11 +8,11 @@
 				<!-- Load WinUI resources -->
 				<XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
 
-				<MaterialColors xmlns="using:Uno.Material"
+				<MaterialColorsV1 xmlns="using:Uno.Material"
 								OverrideSource="ms-appx:///ColorPaletteOverride.xaml" />
 				<MaterialFonts xmlns="using:Uno.Material"
 							   OverrideSource="ms-appx:///MaterialFontsOverride.xaml" />
-				<MaterialResources xmlns="using:Uno.Material" />
+				<MaterialResourcesV1 xmlns="using:Uno.Material" />
 
 				<MaterialColorsV2 xmlns="using:Uno.Material" />
 				<MaterialFonts xmlns="using:Uno.Material"
